Fix market price size from uint32_t to uint64_t

+++ src/creatures/creatures_definitions.hpp
@@ -672,7 +672,7 @@ struct RecentPvPKillEntry {
 };
 
 struct MarketOffer {
-	uint32_t price;
+	uint64_t price;
 	uint32_t timestamp;
 	uint16_t amount;
 	uint16_t counter;
@@ -696,7 +696,7 @@ struct MarketOfferEx {
 	uint32_t id;
 	uint32_t playerId;
 	uint32_t timestamp;
-	uint32_t price;
+	uint64_t price;
 	uint16_t amount;
 	uint16_t counter;
 	uint16_t itemId;
@@ -706,7 +706,7 @@ struct MarketOfferEx {
 
 struct HistoryMarketOffer {
 	uint32_t timestamp;
-	uint32_t price;
+	uint64_t price;
 	uint16_t itemId;
 	uint16_t amount;
 	MarketOfferState_t state;
